// boot-network-ui/openapi/model/registrationRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegistrationRequest { 
    firstName: string;
    lastName: string;
    email: string;
    password: string;
}